Fix documentation regarding how to import/require TscWatchClient from 'tsc-watch/client'
Fixes #156

`export default` is very discouraged in ES6 world:
- https://rajeshnaroth.medium.com/avoid-es6-default-exports-a24142978a7a
- https://blog.piotrnalepa.pl/2020/06/26/default-exports-vs-named-exports/
- https://www.reddit.com/r/javascript/comments/9amkc6/export_default_considered_harmful/

In fact while trying to do it I've faced typical problems because there is some ugly trick to be done for `export default` to not export an Object with a `default` field into it and so on. So in this PR I just fix the documentation in README and in the example script. IMHO this is better since:
- If we change the way `TscWatchClient` must be imported it would mean a compability breaking change.
- `tsc-watch` changed this in 6.0.0 version, which is a mayor update so it's ok to break backwards compatibility. But breaking it again in 6.0.1 is not that good.
- As stated above, dealing with `default export` in ES6 may lead to many problems that we can just avoid by not changing it.

NOTE: Those formatting changes in README are due my text editor. Hope they are not a problem.